src/test/java/com/leetcode/dp/medium/FourKeysKeyboard_651Test.java
package com.leetcode.dp.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class FourKeysKeyboard_651Test {

  private final FourKeysKeyboard_651 solution = new FourKeysKeyboard_651();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(3, 3), arguments(7, 9), arguments(50, 1327104));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void maxA(int n, int expectedResult) {
    int actualResult = solution.maxA(n);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
